fixed process.env
