+++ apps/dashboard/src/app/(main)/(protected)/layout.tsx
@@ -2,7 +2,7 @@
 
 import Loading from "@/app/loading";
 import { useStackApp, useUser } from "@stackframe/stack";
-import { getPublicEnvVar } from '@stackframe/stack-shared/dist/utils/env';
+import { getPublicEnvVar } from '@/lib/env';
 import { runAsynchronouslyWithAlert } from "@stackframe/stack-shared/dist/utils/promises";
 import { useEffect } from "react";
