+++ apps/dashboard/src/app/layout.tsx
@@ -3,11 +3,11 @@ import { RouterProvider } from '@/components/router';
 import { SiteLoadingIndicatorDisplay } from '@/components/site-loading-indicator';
 import { StyleLink } from '@/components/style-link';
 import { ThemeProvider } from '@/components/theme-provider';
-import { cn } from '@/lib/utils';
+import { getPublicEnvVar } from '@/lib/env';
 import { stackServerApp } from '@/stack';
 import { StackProvider, StackTheme } from '@stackframe/stack';
-import { getEnvVariable, getNodeEnvironment, getPublicEnvVar } from '@stackframe/stack-shared/dist/utils/env';
-import { Toaster } from '@stackframe/stack-ui';
+import { getEnvVariable, getNodeEnvironment } from '@stackframe/stack-shared/dist/utils/env';
+import { Toaster, cn } from '@stackframe/stack-ui';
 import { Analytics } from "@vercel/analytics/react";
 import { SpeedInsights } from "@vercel/speed-insights/next";
 import { GeistMono } from "geist/font/mono";
